refactor(settings): update routes
