+++ ui/apps/settings-app/src/routes.ts
@@ -1,6 +1,6 @@
 export const HOME = 'Home';
 export const PRIVACY = 'Privacy';
-export const APPEARANCE = 'Appearance';
+export const THEME = 'Theme';
 export const GENERAL = 'General';
 export const APPS = 'Apps';
 export const PLUGINS = 'Plugins';
@@ -10,6 +10,6 @@ export const rootRoute = '/settings';
 export default {
   [HOME]: `${rootRoute}`,
   [PRIVACY]: `${rootRoute}/privacy`,
-  [APPEARANCE]: `${rootRoute}/appearance`,
+  [THEME]: `${rootRoute}/theme`,
   [APPS]: `${rootRoute}/apps`,
 };
